simplify the asm code
one instruction to set the global interrupt mode.

Notes:
the timer1 interrupt can only trigger one time,
re-download the bin into flash and run it again,
it will not be triggered.
i have to power off the whole system each time.
this is the bug left for future to debug.

Signed-off-by: Zhigang Wu <wzgpeter@163.com>